SplitAnnotationに名前を変更
