--- src/components/Annotation/index.tsx
@@ -1,1 +1,0 @@
-export { default } from "./Annotation";
