+++ src/components/index.ts
@@ -2,8 +2,8 @@
 export { default as ActionButton } from "./ActionButton";
 export * from "./ActionButton";
 
-export { default as Annotation } from "./Annotation";
-export * from "./Annotation";
+export { default as SplitAnnotation } from "./SplitAnnotation";
+export * from "./SplitAnnotation";
 
 export { default as Backdrop } from "./Backdrop";
 export * from "./Backdrop";
